refactor: split tests into files by role
