+++ server/internal/app/expenses_test.go
@@ -344,262 +344,3 @@ func TestGetAllExpenses(t *testing.T) {
 	})
 
 }
-
-func TestCreateUser(t *testing.T) {
-	store := StubExpenseStore{}
-	app := New(&store, &StubOauthConfig{}, &StubSessionManager{}, "", &StubImageStore{})
-
-	user := TestSeanUser
-	userJSON, err := json.Marshal(user)
-	if err != nil {
-		t.Fatalf(err.Error())
-	}
-
-	request := NewCreateUserRequest(userJSON)
-	response := httptest.NewRecorder()
-
-	handler := http.HandlerFunc(app.CreateUser)
-	handler.ServeHTTP(response, request)
-
-	assert.Equal(t, http.StatusAccepted, response.Code)
-	assert.Len(t, store.users, 1)
-	assert.Contains(t, store.users, user)
-}
-
-func TestListUsers(t *testing.T) {
-	store := StubExpenseStore{users: []User{TestSeanUser, TestTomomiUser}}
-	app := New(&store, &StubOauthConfig{}, &StubSessionManager{}, "", &StubImageStore{})
-
-	request := NewGetAllUsersRequest()
-	response := httptest.NewRecorder()
-
-	handler := http.HandlerFunc(app.ListUsers)
-	handler.ServeHTTP(response, request)
-
-	var got []User
-	err := json.NewDecoder(response.Body).Decode(&got)
-	if err != nil {
-		t.Fatalf("error parsing response from server %q into slice of Expenses, '%v'", response.Body, err)
-	}
-
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Len(t, got, len(store.users))
-	assert.ElementsMatch(t, got, store.users)
-}
-
-func TestOauthLogin(t *testing.T) {
-	store := StubExpenseStore{}
-	oauth := StubOauthConfig{}
-	app := New(&store, &oauth, &StubSessionManager{}, "", &StubImageStore{})
-
-	request, err := http.NewRequest(http.MethodGet, "/api/v1/login_google", nil)
-	if err != nil {
-		t.Fatalf("request could not be created, %v", err)
-	}
-	response := httptest.NewRecorder()
-
-	handler := http.HandlerFunc(app.OauthLogin)
-	handler.ServeHTTP(response, request)
-
-	assert.Equal(t, http.StatusTemporaryRedirect, response.Code)
-	// are these even good assertions to have?
-	expectedURL := fmt.Sprintf("/api/v1/%s", oauthProviderMockURL)
-	assert.Equal(t, expectedURL, response.Header().Get("Location"))
-	// assert AuthCodeURL was called
-	assert.Len(t, oauth.AuthCodeURLCalls, 1)
-}
-
-func TestOauthCallback(t *testing.T) {
-	t.Run("creates a user when user doesn't exist yet and creates a new session with the user", func(t *testing.T) {
-		store := StubExpenseStore{users: []User{}}
-		oauth := StubOauthConfig{}
-		sessions := StubSessionManager{}
-		frontend := "http://a.test"
-		app := New(&store, &oauth, &sessions, frontend, &StubImageStore{})
-
-		request := NewGoogleCallbackRequest()
-		response := httptest.NewRecorder()
-
-		handler := http.HandlerFunc(app.OauthCallback)
-		handler.ServeHTTP(response, request)
-
-		// expect a new user to be added to the store, GetInfoAndGenerateUser has been stubbed to generate TestSeanUser
-		expected := []User{TestSeanUser}
-		assert.Len(t, store.users, 1)
-		assert.ElementsMatch(t, expected, store.users)
-
-		assert.Len(t, sessions.saveCalls, 1)
-		assert.Equal(t, sessions.saveCalls[0], TestSeanUser.ID)
-
-		// get routed to the base page for now
-		url, err := response.Result().Location()
-		if err != nil {
-			t.Fatalf("url couldn't be found: %v", err)
-		}
-		assert.Equal(t, frontend, url.String())
-	})
-
-	t.Run("doesn't create a new user when the user already exists, and saves the session with the user in the context", func(t *testing.T) {
-		store := StubExpenseStore{users: []User{TestSeanUser}}
-		oauth := StubOauthConfig{}
-		sessions := StubSessionManager{}
-		frontend := "http://another.test"
-		app := New(&store, &oauth, &sessions, frontend, &StubImageStore{})
-
-		request := NewGoogleCallbackRequest()
-		response := httptest.NewRecorder()
-
-		handler := http.HandlerFunc(app.OauthCallback)
-		handler.ServeHTTP(response, request)
-
-		expected := []User{TestSeanUser}
-		assert.Len(t, store.users, 1)
-		assert.ElementsMatch(t, expected, store.users)
-
-		assert.Len(t, sessions.saveCalls, 1)
-		// the callback will add a context of the appropriate user id
-		assert.Equal(t, sessions.saveCalls[0], TestSeanUser.ID)
-
-		// expect to get routed to the main welcome page
-		url, err := response.Result().Location()
-		if err != nil {
-			t.Fatalf("url couldn't be found: %v", err)
-		}
-		assert.Equal(t, frontend, url.String())
-	})
-}
-
-func TestVerifyUser(t *testing.T) {
-	t.Run("returns a 401 response when the user is not authorized", func(t *testing.T) {
-		store := StubExpenseStore{}
-		oauth := StubOauthConfig{}
-		a := New(&store, &oauth, &StubSessionManager{}, "", &StubImageStore{})
-
-		request := NewGetAllExpensesRequest()
-		response := httptest.NewRecorder()
-
-		handler := a.VerifyUser(http.HandlerFunc(a.GetAllExpenses))
-		handler.ServeHTTP(response, request)
-
-		assert.Equal(t, http.StatusUnauthorized, response.Code)
-	})
-
-	t.Run("returns a 200 response when the user is authorized, and passes the request with the user ID in the context to the appropriate route", func(t *testing.T) {
-		store := StubExpenseStore{expenses: map[string]Expense{"1": TestSeanExpense}}
-		oauth := StubOauthConfig{}
-		a := New(&store, &oauth, &StubSessionManager{}, "", &StubImageStore{})
-
-		request := NewGetAllExpensesRequest()
-		// simulate a cookie session storage here
-		request.AddCookie(&ValidCookie)
-		response := httptest.NewRecorder()
-
-		handler := a.VerifyUser(http.HandlerFunc(a.GetAllExpenses))
-		handler.ServeHTTP(response, request)
-
-		assert.Equal(t, http.StatusOK, response.Code)
-
-		var got []Expense
-		err := json.NewDecoder(response.Body).Decode(&got)
-		if err != nil {
-			t.Fatalf("error parsing response from server %q into slice of Expenses, '%v'", response.Body, err)
-		}
-
-		assert.ElementsMatch(t, got, []Expense{TestSeanExpense})
-	})
-}
-
-func TestGetUserByID(t *testing.T) {
-	t.Run("returns a users details if the user exists", func(t *testing.T) {
-		store := StubExpenseStore{users: []User{TestSeanUser}}
-		oauth := StubOauthConfig{}
-		a := New(&store, &oauth, &StubSessionManager{}, "", &StubImageStore{})
-
-		request := NewGetUserRequest(TestSeanUser.ID)
-		response := httptest.NewRecorder()
-
-		handler := http.HandlerFunc(a.GetUser)
-		handler.ServeHTTP(response, request)
-
-		var got User
-		err := json.NewDecoder(response.Body).Decode(&got)
-		if err != nil {
-			t.Fatalf("error parsing response from server %q into User struct, '%v'", response.Body, err)
-		}
-
-		assert.Equal(t, jsonContentType, response.Result().Header.Get("content-type"))
-		assert.Equal(t, http.StatusOK, response.Code)
-		assert.Equal(t, TestSeanUser, got)
-	})
-}
-
-func TestGetSelf(t *testing.T) {
-	t.Run("returns the user details from the stored session if the user exists", func(t *testing.T) {
-		store := StubExpenseStore{users: []User{TestSeanUser}}
-		oauth := StubOauthConfig{}
-		a := New(&store, &oauth, &StubSessionManager{}, "", &StubImageStore{})
-
-		request := NewGetSelfRequest()
-		// add the user into the request cookie
-		request.AddCookie(&ValidCookie)
-		response := httptest.NewRecorder()
-
-		handler := http.HandlerFunc(a.GetSelf)
-		handler.ServeHTTP(response, request)
-
-		// decode the response
-		var got User
-		err := json.NewDecoder(response.Body).Decode(&got)
-		if err != nil {
-			t.Fatalf("error parsing response from server %q into slice of Expenses, '%v'", response.Body, err)
-		}
-
-		assert.Equal(t, jsonContentType, response.Result().Header.Get("content-type"))
-		assert.Equal(t, http.StatusOK, response.Code)
-		assert.Equal(t, TestSeanUser, got)
-	})
-
-	t.Run("returns a 404 if the user does not exist", func(t *testing.T) {
-		store := StubExpenseStore{users: []User{TestSeanUser}}
-		oauth := StubOauthConfig{}
-		a := New(&store, &oauth, &StubSessionManager{}, "", &StubImageStore{})
-
-		request := NewGetSelfRequest()
-		// add the user into the request cookie
-		request.AddCookie(&http.Cookie{
-			Name:  ValidCookie.Name,
-			Value: "non-existent-user",
-		})
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(a.GetSelf)
-		handler.ServeHTTP(response, request)
-
-		assert.Equal(t, http.StatusNotFound, response.Code)
-		assert.NotEqual(t, jsonContentType, response.Result().Header.Get("content-type"))
-	})
-}
-
-func TestLogOut(t *testing.T) {
-	t.Run("session manager calls remove", func(t *testing.T) {
-		store := StubExpenseStore{users: []User{TestSeanUser}}
-		oauth := StubOauthConfig{}
-		sessions := StubSessionManager{}
-		frontend := "http://test.base"
-		a := New(&store, &oauth, &sessions, frontend, &StubImageStore{})
-
-		request, _ := http.NewRequest(http.MethodGet, "/api/v1/logout", nil)
-		response := httptest.NewRecorder()
-
-		handler := http.HandlerFunc(a.LogOut)
-		handler.ServeHTTP(response, request)
-
-		assert.Equal(t, 1, sessions.removeCalls)
-		assert.Equal(t, http.StatusTemporaryRedirect, response.Code)
-
-		url, err := response.Result().Location()
-		if err != nil {
-			t.Fatalf("url couldn't be found: %v", err)
-		}
-		assert.Equal(t, frontend, url.String())
-	})
-}
